requirements.txt
streamlit
pandas
numpy
Pillow
plotly
scikit-learn
xgboost
tensorflow
joblib
requests
python-dotenv
fastapi
uvicorn
pydantic
pickle